+++ presto-datasketches/src/main/java/com/facebook/presto/facebook/tdigest/docs/tdigest.graffle
